chore : category can be created and updated

+++ src/ExpenseTracker.Core/Repositories/Interface/IGenericRepository.cs
@@ -19,8 +19,8 @@ namespace ExpenseTracker.Core.Repositories.Interface
         Task<IList<T>> GetAllAsync();
         IQueryable<T> GetPredicatedQueryable(Expression<Func<T, bool>>? predicate);
         IQueryable<T> GetQueryable();
-        T GetById(long id);
-        Task<T?> GetByIdAsync(long id);
+        T GetById(int id);
+        Task<T?> GetByIdAsync(int id);
 
         Task<bool> CheckIfExistAsync(Expression<Func<T, bool>> predicate);
