+++ src/ExpenseTracker.Core/Services/Implementation/TransactionCategoryService.cs
@@ -44,7 +44,7 @@ namespace ExpenseTracker.Core.Services.Implementation
             Tx.Complete();
         }
 
-        public async Task Delete(long transactionCategoryId)
+        public async Task Delete(int transactionCategoryId)
         {
             var transactionCategoryExists = await _transactionCategoryRepository.CheckIfExistAsync(a=>a.TransactionCategoryId == transactionCategoryId).ConfigureAwait(false);
             if (!transactionCategoryExists) throw new TransactionCategoryNotFoundException(transactionCategoryId);
